util: Ensure ChMask indexes are uint8

+++ pkg/band/band.go
@@ -251,15 +251,15 @@ var usAuBeaconFrequencies = func() [8]uint32 {
 	return freqs
 }()
 
-func chMask16Channels(mask [16]bool, cntl uint8) (map[int]bool, error) {
-	chans := make(map[int]bool, 16)
+func chMask16Channels(mask [16]bool, cntl uint8) (map[uint8]bool, error) {
+	chans := make(map[uint8]bool, 16)
 	switch cntl {
 	case 0:
-		for i := 0; i < 16; i++ {
+		for i := uint8(0); i < 16; i++ {
 			chans[i] = mask[i]
 		}
 	case 6:
-		for i := 0; i < 16; i++ {
+		for i := uint8(0); i < 16; i++ {
 			chans[i] = true
 		}
 	default:
@@ -268,25 +268,25 @@ func chMask16Channels(mask [16]bool, cntl uint8) (map[int]bool, error) {
 	return chans, nil
 }
 
-func chMask72Channels(mask [16]bool, cntl uint8) (map[int]bool, error) {
-	chans := make(map[int]bool, 72)
+func chMask72Channels(mask [16]bool, cntl uint8) (map[uint8]bool, error) {
+	chans := make(map[uint8]bool, 72)
 	switch cntl {
 	case 0, 1, 2, 3, 4:
-		for i := 0; i < 72; i++ {
-			chans[i] = (i >= int(cntl)*16 && i < int(cntl+1)*16) && mask[i%16]
+		for i := uint8(0); i < 72; i++ {
+			chans[i] = (i >= cntl*16 && i < (cntl+1)*16) && mask[i%16]
 		}
 	case 5:
-		for i := 0; i < 64; i++ {
+		for i := uint8(0); i < 64; i++ {
 			chans[i] = mask[i/8]
 		}
-		for i := 64; i < 72; i++ {
+		for i := uint8(64); i < 72; i++ {
 			chans[i] = mask[i-64]
 		}
 	case 6, 7:
-		for i := 0; i < 64; i++ {
+		for i := uint8(0); i < 64; i++ {
 			chans[i] = cntl == 6
 		}
-		for i := 64; i < 72; i++ {
+		for i := uint8(64); i < 72; i++ {
 			chans[i] = mask[i-64]
 		}
 	default:
@@ -295,15 +295,15 @@ func chMask72Channels(mask [16]bool, cntl uint8) (map[int]bool, error) {
 	return chans, nil
 }
 
-func chMask96Channels(mask [16]bool, cntl uint8) (map[int]bool, error) {
-	chans := make(map[int]bool, 96)
+func chMask96Channels(mask [16]bool, cntl uint8) (map[uint8]bool, error) {
+	chans := make(map[uint8]bool, 96)
 	switch cntl {
 	case 0, 1, 2, 3, 4, 5:
-		for i := 0; i < 96; i++ {
-			chans[i] = (i >= int(cntl)*16 && i < int(cntl+1)*16) && mask[i%16]
+		for i := uint8(0); i < 96; i++ {
+			chans[i] = (i >= cntl*16 && i < (cntl+1)*16) && mask[i%16]
 		}
 	case 6:
-		for i := 0; i < 96; i++ {
+		for i := uint8(0); i < 96; i++ {
 			chans[i] = true
 		}
 	default:
